backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
